f '[Vagrant] Add gitian-based, vagrant VM build system to enable building on any operating system platform.'
Execute `make -C depends download` from within the gitian host linux environment, as the choice of which packages to download is dependent on the operating system within which it is executed. On macOS and Windows the wrong set of packages are downloaded, since the gitian build takes place within a linux environment.